+++ src/main/java/com/example/demo/dto/ingredientdtos/CreateOrUpdateIngredientDTO.java
@@ -1,7 +1,8 @@
 package com.example.demo.dto.ingredientdtos;
 
+import java.math.BigDecimal;
 
 public class CreateOrUpdateIngredientDTO {
     public String name;
-    public Double price;
+    public BigDecimal price;
 }
