+++ src/main/java/com/example/demo/dto/pizzadtos/PizzaDTO.java
@@ -1,6 +1,7 @@
 package com.example.demo.dto.pizzadtos;
 
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
@@ -9,7 +10,7 @@ import java.util.UUID;
 public class PizzaDTO {
     public UUID id;
     public String name;
-    public Double price;
+    public BigDecimal price;
     public String description;
     public Set<PizzaIngredientDTO> ingredients = new HashSet<PizzaIngredientDTO>();
 }
